refactor models to use pydantic, modify client

+++ zep_python/zep_client.py
@@ -1,12 +1,13 @@
 from __future__ import annotations
 
+import asyncio
 from types import TracebackType
 from typing import List, Optional, Type
 
 import httpx
 
 from zep_python.exceptions import APIError, NotFoundError
-from zep_python.models import Memory, SearchPayload, SearchResult
+from zep_python.models import Memory, Message, SearchPayload, SearchResult, Summary
 from zep_python.utils import sync
 
 API_BASEURL = "/api/v1"
@@ -54,19 +55,30 @@ class ZepClient:
     async def __aenter__(self) -> "ZepClient":
         return self
 
-    # Asynchronous context manager exit point
+    # Sync context manager exit point
     async def __aexit__(
         self,
         exc_type: Type[Exception],
         exc_val: Exception,
         exc_tb: TracebackType,
     ) -> None:
-        await self.close()
+        await self.aclose()
+
+    # Asynchronous context manager entry point
+    def __enter__(self) -> "ZepClient":
+        return self
+
+    # Sync context manager exit point
+    def __exit__(
+        self,
+        exc_type: Type[Exception],
+        exc_val: Exception,
+        exc_tb: TracebackType,
+    ) -> None:
+        self.close()
 
     @sync
-    def get_memory(
-        self, session_id: str, lastn: Optional[int] = None
-    ) -> List[Memory]:
+    def get_memory(self, session_id: str, lastn: Optional[int] = None) -> List[Memory]:
         """
         Retrieve memory for the specified session. This method is a synchronous wrapper
         for the asynchronous method `aget_memory`.
@@ -91,6 +103,9 @@ class ZepClient:
         """
         # we've wrapped the function in a decorator that will run it synchronously.
         # ignore the type error.
+        if session_id is None or session_id == "":
+            raise ValueError("session_id must be provided")
+
         return self.aget_memory(session_id, lastn)  # type: ignore
 
     async def aget_memory(
@@ -117,6 +132,9 @@ class ZepClient:
         APIError
             If the API response format is unexpected.
         """
+        if session_id is None or session_id == "":
+            raise ValueError("session_id must be provided")
+
         url = f"{self.base_url}{API_BASEURL}/sessions/{session_id}/memory"
         params = (
             {"lastn": lastn} if lastn is not None else {}
@@ -130,23 +148,28 @@ class ZepClient:
             raise APIError(f"Unexpected status code: {response.status_code}")
 
         response_data = response.json()
+        messages: List[Message]
+        try:
+            messages = [
+                Message.parse_obj(m) for m in response_data.get("messages", None)
+            ]
+            if len(messages) == 0:
+                raise ValueError("Messages can't be empty")
+        except (TypeError, ValueError) as e:
+            raise APIError("Unexpected response format from the API") from e
 
-        # Check if the response contains the expected field 'messages'.
-        if "messages" in response_data:
-            # If 'messages' is None, return an empty list.
-            if response_data["messages"] is None:
-                return []
-            # Create a Memory instance using the 'messages' field from the response.
-            memory = Memory(
-                messages=response_data["messages"],
-                # Add the 'summary' field if it is present in the response.
-                summary=response_data.get("summary", None),
-                # Add any other fields from the response that are relevant to the
-                # Memory class.
-            )
-            return [memory]
-        else:
-            raise APIError("Unexpected response format from the API")
+        summary: Optional[Summary] = None
+        if response_data.get("summary", None) is not None:
+            summary = Summary.parse_obj(response_data["summary"])
+
+        memory = Memory(
+            messages=messages,
+            # Add the 'summary' field if it is present in the response.
+            summary=summary,
+            # Add any other fields from the response that are relevant to the
+            # Memory class.
+        )
+        return [memory]
 
     @sync
     def add_memory(self, session_id: str, memory_messages: Memory) -> str:
@@ -173,11 +196,12 @@ class ZepClient:
         """
         # we've wrapped the function in a decorator that will run it synchronously.
         # ignore the type error.
+        if session_id is None or session_id == "":
+            raise ValueError("session_id must be provided")
+
         return self.aadd_memory(session_id, memory_messages)  # type: ignore
 
-    async def aadd_memory(
-        self, session_id: str, memory_messages: Memory
-    ) -> str:
+    async def aadd_memory(self, session_id: str, memory_messages: Memory) -> str:
         """
         Asynchronously add memory to the specified session.
 
@@ -198,9 +222,12 @@ class ZepClient:
         APIError
             If the API response format is unexpected.
         """
+        if session_id is None or session_id == "":
+            raise ValueError("session_id must be provided")
+
         response = await self.client.post(
             f"{self.base_url}{API_BASEURL}/sessions/{session_id}/memory",
-            json=memory_messages.to_dict(),
+            json=memory_messages.dict(exclude_none=True),
         )
         if response.status_code != 200:
             raise APIError(f"Unexpected status code: {response.status_code}")
@@ -230,6 +257,9 @@ class ZepClient:
         """
         # we've wrapped the function in a decorator that will run it synchronously.
         # ignore the type error.
+        if session_id is None or session_id == "":
+            raise ValueError("session_id must be provided")
+
         return self.adelete_memory(session_id)  # type: ignore
 
     async def adelete_memory(self, session_id: str) -> str:
@@ -251,6 +281,9 @@ class ZepClient:
         APIError
             If the API response format is unexpected.
         """
+        if session_id is None or session_id == "":
+            raise ValueError("session_id must be provided")
+
         response = await self.client.delete(
             f"{self.base_url}{API_BASEURL}/sessions/{session_id}/memory"
         )
@@ -293,6 +326,9 @@ class ZepClient:
         """
         # we've wrapped the function in a decorator that will run it synchronously.
         # ignore the type error.
+        if session_id is None or session_id == "":
+            raise ValueError("session_id must be provided")
+
         return self.asearch_memory(session_id, search_payload, limit)  # type: ignore
 
     async def asearch_memory(
@@ -323,21 +359,25 @@ class ZepClient:
         APIError
             If the API response format is unexpected.
         """
+        if session_id is None or session_id == "":
+            raise ValueError("session_id must be provided")
+
+        if search_payload is None:
+            raise ValueError("search_payload must be provided")
+
         params = {"limit": limit} if limit is not None else {}
         response = await self.client.post(
             f"{self.base_url}{API_BASEURL}/sessions/{session_id}/search",
-            json=search_payload.__dict__,
+            json=search_payload.dict(),
             params=params,
         )
         if response.status_code == 404:
             raise NotFoundError("No query results found")
         if response.status_code != 200:
             raise APIError(f"Unexpected status code: {response.status_code}")
-        return [
-            SearchResult(**search_result) for search_result in response.json()
-        ]
+        return [SearchResult(**search_result) for search_result in response.json()]
 
-    async def close(self) -> None:
+    async def aclose(self) -> None:
         """
         Asynchronously close the HTTP client.
 
@@ -345,3 +385,12 @@ class ZepClient:
         release resources.
         """
         await self.client.aclose()
+
+    def close(self) -> None:
+        """
+        Close the HTTP client.
+
+        [Optional] This method may be called when the ZepClient is no longer needed to
+        release resources.
+        """
+        asyncio.run(self.client.aclose())
